backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
python-decouple
whitenoise
